// src/app/models/vertical-data-collection/v-form-constructor/v-form-builder/active-section.model.ts
export interface IActiveSections {
  formFields: IActiveSection;
  documentsForms: IActiveSection;
  consent: IActiveSection;
  tuition: IActiveSection;
  paymentSettings: IActiveSection;
  termsConditionals: IActiveSection;
  formPayment: IActiveSection;
}

export interface IActiveSection {
  isActive: boolean;
  showSideInfo: boolean;
}

export const activeSectionsDefault: IActiveSections = {
  formFields: {isActive: false, showSideInfo: false},
  documentsForms: {isActive: false, showSideInfo: false},
  consent: {isActive: false, showSideInfo: false},
  tuition: {isActive: false, showSideInfo: false},
  paymentSettings: {isActive: false, showSideInfo: false},
  termsConditionals: {isActive: false, showSideInfo: false},
  formPayment: {isActive: false, showSideInfo: false}
};